requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
graphene==3.2.2
graphene-django==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
Markdown==3.4.3
mysqlclient==2.1.1
promise==2.3
protobuf==4.23.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
